fix: separate image mod from core
