+++ cradle/linux/src/main.rs
@@ -1,6 +1,5 @@
 
 #[macro_use] extern crate log;
-#[macro_use] extern crate peridot_derive;
 
 use std::fs::File;
 use std::path::PathBuf;
