+++ image/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "peridot-image"
+version = "0.1.0"
+description = "Peridot Engine Image Management"
+authors = ["S.Percentage <Syn.Tri.Naga@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+peridot = { path = ".." }
+image = "*"
